fixed some -Wshorten-64-to-32 warnings

+++ src/dft_func/XCIntegratorForGGA.cpp
@@ -197,7 +197,7 @@ integrateVxcFockForGgaClosedShell(const CMolecule&                  molecule,
 
             // go through GTO blocks
 
-            for (size_t i_block = 0, idx = 0; i_block < n_gto_blocks; i_block++)
+            for (int i_block = 0, idx = 0; i_block < static_cast<int>(n_gto_blocks); i_block++)
             {
                 const auto& gto_block = ptr_gto_blocks[i_block];
 
@@ -500,7 +500,7 @@ integrateVxcFockForGgaOpenShell(const CMolecule&                  molecule,
 
             // go through GTO blocks
 
-            for (size_t i_block = 0, idx = 0; i_block < n_gto_blocks; i_block++)
+            for (int i_block = 0, idx = 0; i_block < static_cast<int>(n_gto_blocks); i_block++)
             {
                 const auto& gto_block = ptr_gto_blocks[i_block];
 
@@ -793,7 +793,7 @@ integrateFxcFockForGgaClosedShell(const std::vector<double*>&       aoFockPointe
 
             std::vector<CDenseMatrix> rw_sub_dens_mat_vec(n_rw_densities);
 
-            for (size_t idensity = 0; idensity < n_rw_densities; idensity++)
+            for (int idensity = 0; idensity < static_cast<int>(n_rw_densities); idensity++)
             {
                 rw_sub_dens_mat_vec[idensity] = dftsubmat::getSubDensityMatrix(rwDensityPointers[idensity], aoinds, naos);
             }
@@ -819,7 +819,7 @@ integrateFxcFockForGgaClosedShell(const std::vector<double*>&       aoFockPointe
 
             // go through GTO blocks
 
-            for (size_t i_block = 0, idx = 0; i_block < n_gto_blocks; i_block++)
+            for (int i_block = 0, idx = 0; i_block < static_cast<int>(n_gto_blocks); i_block++)
             {
                 const auto& gto_block = ptr_gto_blocks[i_block];
 
@@ -905,7 +905,7 @@ integrateFxcFockForGgaClosedShell(const std::vector<double*>&       aoFockPointe
 
             // go through rhow density matrices
 
-            for (size_t idensity = 0; idensity < n_rw_densities; idensity++)
+            for (int idensity = 0; idensity < static_cast<int>(n_rw_densities); idensity++)
             {
                 omptimers[thread_id].start("Generate density grid");
 
@@ -1175,7 +1175,7 @@ integrateFxcFockForGgaOpenShell(const std::vector<double*>&       aoFockPointers
             std::vector<CDenseMatrix> rw_sub_dens_mat_vec_a(n_rw_densities);
             std::vector<CDenseMatrix> rw_sub_dens_mat_vec_b(n_rw_densities);
 
-            for (size_t idensity = 0; idensity < n_rw_densities; idensity++)
+            for (int idensity = 0; idensity < static_cast<int>(n_rw_densities); idensity++)
             {
                 rw_sub_dens_mat_vec_a[idensity] = dftsubmat::getSubDensityMatrix(rwDensityPointers[idensity * 2 + 0], aoinds, naos);
                 rw_sub_dens_mat_vec_b[idensity] = dftsubmat::getSubDensityMatrix(rwDensityPointers[idensity * 2 + 1], aoinds, naos);
@@ -1202,7 +1202,7 @@ integrateFxcFockForGgaOpenShell(const std::vector<double*>&       aoFockPointers
 
             // go through GTO blocks
 
-            for (size_t i_block = 0, idx = 0; i_block < n_gto_blocks; i_block++)
+            for (int i_block = 0, idx = 0; i_block < static_cast<int>(n_gto_blocks); i_block++)
             {
                 const auto& gto_block = ptr_gto_blocks[i_block];
 
@@ -1288,7 +1288,7 @@ integrateFxcFockForGgaOpenShell(const std::vector<double*>&       aoFockPointers
 
             // go through rhow density matrices
 
-            for (size_t idensity = 0; idensity < n_rw_densities; idensity++)
+            for (int idensity = 0; idensity < static_cast<int>(n_rw_densities); idensity++)
             {
                 omptimers[thread_id].start("Generate density grid");
 
@@ -1636,7 +1636,7 @@ integrateKxcFockForGgaClosedShell(const std::vector<double*>& aoFockPointers,
 
             // go through GTO blocks
 
-            for (size_t i_block = 0, idx = 0; i_block < n_gto_blocks; i_block++)
+            for (int i_block = 0, idx = 0; i_block < static_cast<int>(n_gto_blocks); i_block++)
             {
                 const auto& gto_block = ptr_gto_blocks[i_block];
 
@@ -1983,7 +1983,7 @@ integrateKxcLxcFockForGgaClosedShell(const std::vector<double*>& aoFockPointers,
 
             // go through GTO blocks
 
-            for (size_t i_block = 0, idx = 0; i_block < n_gto_blocks; i_block++)
+            for (int i_block = 0, idx = 0; i_block < static_cast<int>(n_gto_blocks); i_block++)
             {
                 const auto& gto_block = ptr_gto_blocks[i_block];
