+++ src/dft_func/XCIntegratorForLDA.cpp
@@ -194,7 +194,7 @@ integrateVxcFockForLdaClosedShell(const CMolecule&                  molecule,
 
             // go through GTO blocks
 
-            for (size_t i_block = 0, idx = 0; i_block < n_gto_blocks; i_block++)
+            for (int i_block = 0, idx = 0; i_block < static_cast<int>(n_gto_blocks); i_block++)
             {
                 const auto& gto_block = ptr_gto_blocks[i_block];
 
@@ -465,7 +465,7 @@ integrateVxcFockForLdaOpenShell(const CMolecule&                  molecule,
 
             // go through GTO blocks
 
-            for (size_t i_block = 0, idx = 0; i_block < n_gto_blocks; i_block++)
+            for (int i_block = 0, idx = 0; i_block < static_cast<int>(n_gto_blocks); i_block++)
             {
                 const auto& gto_block = ptr_gto_blocks[i_block];
 
@@ -721,7 +721,7 @@ integrateFxcFockForLdaClosedShell(const std::vector<double*>&       aoFockPointe
 
             std::vector<CDenseMatrix> rw_sub_dens_mat_vec(n_rw_densities);
 
-            for (size_t idensity = 0; idensity < n_rw_densities; idensity++)
+            for (int idensity = 0; idensity < static_cast<int>(n_rw_densities); idensity++)
             {
                 rw_sub_dens_mat_vec[idensity] = dftsubmat::getSubDensityMatrix(rwDensityPointers[idensity], aoinds, naos);
             }
@@ -744,7 +744,7 @@ integrateFxcFockForLdaClosedShell(const std::vector<double*>&       aoFockPointe
 
             // go through GTO blocks
 
-            for (size_t i_block = 0, idx = 0; i_block < n_gto_blocks; i_block++)
+            for (int i_block = 0, idx = 0; i_block < static_cast<int>(n_gto_blocks); i_block++)
             {
                 const auto& gto_block = ptr_gto_blocks[i_block];
 
@@ -801,7 +801,7 @@ integrateFxcFockForLdaClosedShell(const std::vector<double*>&       aoFockPointe
 
             // go through rhow density matrices
 
-            for (size_t idensity = 0; idensity < n_rw_densities; idensity++)
+            for (int idensity = 0; idensity < static_cast<int>(n_rw_densities); idensity++)
             {
                 omptimers[thread_id].start("Generate density grid");
 
@@ -995,7 +995,7 @@ integrateFxcFockForLdaOpenShell(const std::vector<double*>&       aoFockPointers
             std::vector<CDenseMatrix> rw_sub_dens_mat_vec_a(n_rw_densities);
             std::vector<CDenseMatrix> rw_sub_dens_mat_vec_b(n_rw_densities);
 
-            for (size_t idensity = 0; idensity < n_rw_densities; idensity++)
+            for (int idensity = 0; idensity < static_cast<int>(n_rw_densities); idensity++)
             {
                 rw_sub_dens_mat_vec_a[idensity] = dftsubmat::getSubDensityMatrix(rwDensityPointers[idensity * 2 + 0], aoinds, naos);
                 rw_sub_dens_mat_vec_b[idensity] = dftsubmat::getSubDensityMatrix(rwDensityPointers[idensity * 2 + 1], aoinds, naos);
@@ -1019,7 +1019,7 @@ integrateFxcFockForLdaOpenShell(const std::vector<double*>&       aoFockPointers
 
             // go through GTO blocks
 
-            for (size_t i_block = 0, idx = 0; i_block < n_gto_blocks; i_block++)
+            for (int i_block = 0, idx = 0; i_block < static_cast<int>(n_gto_blocks); i_block++)
             {
                 const auto& gto_block = ptr_gto_blocks[i_block];
 
@@ -1076,7 +1076,7 @@ integrateFxcFockForLdaOpenShell(const std::vector<double*>&       aoFockPointers
 
             // go through rhow density matrices
 
-            for (size_t idensity = 0; idensity < n_rw_densities; idensity++)
+            for (int idensity = 0; idensity < static_cast<int>(n_rw_densities); idensity++)
             {
                 omptimers[thread_id].start("Generate density grid");
 
@@ -1298,7 +1298,7 @@ integrateKxcFockForLdaClosedShell(const std::vector<double*>& aoFockPointers,
 
             // go through GTO blocks
 
-            for (size_t i_block = 0, idx = 0; i_block < n_gto_blocks; i_block++)
+            for (int i_block = 0, idx = 0; i_block < static_cast<int>(n_gto_blocks); i_block++)
             {
                 const auto& gto_block = ptr_gto_blocks[i_block];
 
@@ -1569,7 +1569,7 @@ integrateKxcLxcFockForLdaClosedShell(const std::vector<double*>& aoFockPointers,
 
             // go through GTO blocks
 
-            for (size_t i_block = 0, idx = 0; i_block < n_gto_blocks; i_block++)
+            for (int i_block = 0, idx = 0; i_block < static_cast<int>(n_gto_blocks); i_block++)
             {
                 const auto& gto_block = ptr_gto_blocks[i_block];
